Multidimensional Integration using Vegas+

An R wrapper for the Python numerical integration library vegas. This allows integrals defined as R functions (including using Rcpp and associated packages) to be passed to the python library for computation. Some minimal requirements apply to the functions. See Details.

Usage

vegas(
  f,
  lower,
  upper,
  nitn_warm = 10,
  neval_warm = 1000,
  nitn = 10,
  neval = 1000,
  errTol = 1,
  maxIter = 5,
  seed = 99999,
  extra_args = list()
)

Arguments

f

An R function that takes a matrix and returns a vector. See details and examples.

lower

A vector of lower integration limits for each dimension, e.g. c(-1.,-1.-1)

upper

A vector of upper integration limits for each dimension, e.g. c(1.,1.1)

nitn_warm

Number of iterations for Vegas warm-up.

neval_warm

Number of function evaluations per iteration in warm-up.

nitn

Number of iterations post-warm-up, these results contribute to final estimate and final error.

neval

Number of function evaluations per iteration post-warm-up, these results contribute to final estimate and final error.

errTol

the % error target, default is 1, i.e. error is 1% of current estimated integral value

maxIter

max number of iteration blocks to run to achieve errTol. Each block comprises nitn iterations

seed

random number seed for vegas sampling generating. set for reproducible results. This is vegas' python random number generator not R's.

extra_args

a named list of additional arguments passed to the function f. These must be numeric vectors or matrices. See details.

Details

The vegas Python library implements the 2020 VEGAS+ adaptive Monte Carlo numerical integration algorithm. This is update of the original vegas algorithm from 1978 and by the original author G. P. Lepage, who also maintains the Python library. This library is a thin R wrapper making the key functions accessible from R. For more information on the Vegas algorithm, visit https://vegas.readthedocs.io/ and the VEGAS+ arXiv article is here.

The R function passed defines an integral and its structure and arguments need to meet some minimum requirements:

The first argument passed to the R function from Python vegas will be a numerical matrix of shape [BATCH,Dim], and vegas will expect back a numerical vector of length Dim. Each row in the matrix is a single set of values of the integration variables, e.g. for a 5-D integrand then this matrix must have 5 columns. Reticulate takes care of conversion between R and Python matrices and vectors but the dimensions must be as described.

Additional named arguments can be passed and these should always be matrix() objects, with integers converted to float (or have trailing period added to avoid ambiguity)

If an additional argument is a scalar, say svalue, then use mvalue<-matrix(data=c(svalue,nrow=1), coercing to float first if necessary, and in the R function use mvalue[1].

The warm-up period is mandatory although it can be reduced to only one iteration. The warm-up is the very start of grid adaptation. Warm-up results are discarded as while they do not bias the result there inclusion can be inefficient due to increasing the estimated error.

The python vegas function takes nitn and neval arguments but does not have a tolerance argument. This function runs the python vegas function repeatedly if necessary to reduce the error down to the target errTol level. The repeats are in blocks of nitn and neval and the final estimate and error from these blocks is combined into an overall weighted final estimate and final error. The functions necessary for this weighting and combining are provided in the vegas python library, specifically vegas.ravg().

References

G. P. Lepage, “A New Algorithm for Adaptive Multidimensional Integration” J. Comp. Phys. 27, 192–203 (1978).

G. P. Lepage, “Adaptive multidimensional integration: vegas enhanced” J. Comp. Phys. 439 (2021) 110386.

https://doi.org/10.1016/j.jcp.2021.110386

Examples

if (FALSE) { # \dontrun{
# Use library(mvtnorm) first for 3-D Multivariate Normal Density
library(mvtnorm)
mu<-matrix(c(0.5, -0.2, 0.1),nrow=1) # note matrix
cov<-matrix(data=c(
 1.0, 0.5, 0.2,
 0.5, 1.2, 0.3,
 0.2, 0.3, 0.8),ncol=3,byrow=FALSE)

# use built-in pmvnorm()
lower=c(-0.5,-0.5,-0.5); upper=c(2.,1.,3.)
res_builtin<-pmvnorm(lower=lower,upper=upper,
                    mean=as.numeric(mu), # coercion as needs vector
                    sigma=cov)
print(res_builtin)

## now use vegas to compute same integral
library(vegasr)
## Important - run next line in each R session to ensure python is ready
vegas_initialize()

# the integrand function MUST take a matrix of dimension [BATCH,M] as first
# argument and return a vector of length M (or 1-row matrix).
# Any number of other additional *named* arguments are allowed and should all
# be matrices. See ?vegas_integrate for more details.
myf<-function(x,mu,cov){
  res<-dmvnorm(x,
               mean = mu, # this is a 1-row matrix, dmvnorm accepts this
               sigma=cov)
  return(res)
}

## See help page for descriptions of warm and nitn and neval.
vegas_result<-vegas(f=myf,
                              lower=lower, upper=upper,
                              nitn_warm = 10, neval_warm = 10000,
                              nitn = 10, neval = 10000,
                              errTol=0.1, maxIter=20,seed=99999,
                              extra_args=list(mu=mu, cov=cov))
# extra_args are additional arguments needed for myf
print(vegas_result)

} # }