Compute Bayesian Marginal Posterior

This function returns the estimated marginal posterior density, f(z) at a value of random variable z, where the remaining dimensions/variables are integrated out. See vignette("bayes1", package = "vegasr") and vignette("rcpp", package = "vegasr")for examples and usage.

Usage

vegasBayesPosterior(
  f,
  lower,
  upper,
  nitn_warm = 10,
  neval_warm = 1000,
  nitn = 10,
  neval = 1000,
  errTol = 1,
  maxIter = 5,
  seed = 99999,
  nsearch = 1000,
  log_evidence,
  extra_args = list()
)

Arguments

f

An R function that takes a matrix and returns a vector. See details and examples.

lower

A vector of lower integration limits for each dimension, e.g. c(-1.,-1.-1)

upper

A vector of upper integration limits for each dimension, e.g. c(1.,1.1)

nitn_warm

Number of iterations for Vegas warm-up

neval_warm

Number of function evaluations per iteration in warm-up

nitn

Number of iterations post-warm-up.

neval

Number of function evaluations per iteration post-warm-up.

errTol

the % error target, default is 1, i.e. error is 1% of current estimated integral value

maxIter

max number of iteration blocks to run to achieve errTol. Each block comprises nitn iterations

seed

random number seed for vegas sampling generating. set for reproducible results.

nsearch

number of points to evaluate log_posterior to find approx max value for shiftby. See details.

extra_args

a named list of additional arguments passed to the function f. This must include at least uselog and shiftby arguments which are mandatory for the function f.

Details

The function passed must meet some specific criteria and a range of example functions are included in the package, e.g. see fn_log_post_1, in particular the function deal with integration limits via transformation if necessary. Example functions using Rcpp are also provided, see arma_fn_log_post_1 which uses RcppArmadillo, eigen_fn_log_post_1 which uses Eigen and eigen_fn_log_post_1_par which used Eigen and RcppParallel.

To help avoid numerical underflow the log posterior values are location shifted where a maximum value is subtracted on the log scale. This maximum value is estimated using a simple grid search between the upper and lower bounds. This maximum value does not need to be precise but a higher nsearch value may potentially help in case of NaN or underflow.