Multidimensional Integration using Vegas+

An R wrapper for the Python numerical integration library vegas. This allows integrals defined as R functions to be passed to the python library for computation. Some minimal requirements apply to the functions. See Details.

Usage

vegas_integrate(
  f,
  lower,
  upper,
  nitn_warm = 10,
  neval_warm = 1000,
  nitn = 10,
  neval = 1000,
  errTol = 1,
  maxIter = 5,
  ...
)

Arguments

f

An R function that takes a matrix and returns a vector. See details and examples.

lower

A vector of lower integration limits for each dimension, e.g. c(-1.,-1.-1)

upper

A vector of upper integration limits for each dimension, e.g. c(1.,1.1)

nitn_warm

Number of iterations for Vegas warmup

neval_warm

Number of function evaluations per iteration in warmup

nitn

Number of iterations post-warmup.

neval

Number of function evaluations per iteration post-warmup.

errTol

the % error target, default is 1, i.e. error is 1% of current estimated integral value

maxIter

max number of iteration blocks to run to achieve errTol. Each block comprises nitn iterations.

...

Additional arguments passed to the function f. These must be numeric vectors or matrices. See details.

Details

The vegas Python library implements the 2020 VEGAS+ adaptive Monte Carlo numerical integration algorithm. This is update of the original vegas algorithm from 1978 and by the original author G. P. Lepage, who also maintains the Python library. This library is a thin R wrapper making the key functions accessible from R. For more information on the Vegas algorithm, visit https://vegas.readthedocs.io/ and the VEGAS+ arXiv article is here.

The R function passed defines an integral and its structure and arguments need to meet some minimum requirements:

The first argument passed to the R function from Python vegas will be a numerical matrix of shape [BATCH,Dim], and vegas will expect back a numerical vector of length Dim. Each row in the matrix is a single set of values of the integration variables, e.g. for a 5-D integrand then this matrix must have 5 columns. Reticulate takes care of conversion between R and Python matrices and vectors but the dimensions must be as described.

Additional named arguments can be passed and these should always be matrix() objects, with integers converted to float (or have trailing period added to avoi ambiguity)

If an additional argument is a scalar, say svalue, then use mvalue<-matrix(data=c(svalue,nrow=1), coercing to float first if necessary, and in the R function use mvalue[1].

References

G. P. Lepage, “A New Algorithm for Adaptive Multidimensional Integration” J. Comp. Phys. 27, 192–203 (1978).

G. P. Lepage, “Adaptive multidimensional integration: vegas enhanced” J. Comp. Phys. 439 (2021) 110386.

https://doi.org/10.1016/j.jcp.2021.110386

Examples

if (FALSE) { # \dontrun{
library(vegasr)
vegas_initialize() # only needed once per session

### EXAMPLE 1 - use of additional matrix arguments
library(mvtnorm)
mu<-matrix(c(0.5, -0.2, 0.1),nrow=1)
cov<-matrix(data=c(
                1.0, 0.5, 0.2,
                0.5, 1.2, 0.3,
                0.2, 0.3, 0.8),ncol=3,byrow=FALSE)

myf<-function(x,mu,cov){
 res<-dmvnorm(x,mean = mu,sigma=cov)
 return(res)
 }

result<-vegas_integrate(f=myf,
                        lower=c(-5,-5,-5), upper=c(5,5,5),
                        nitn_warm = 10, neval_warm = 10000,
                        nitn = 10, neval = 10000,
                        errTol=0.1,maxIter=20,
                        mu=mu,cov=cov)
print(result)
# mean should be very close to 0.999 and error close to 0.005

########### Estimate 5-D integral with known solution to check correctness
# define integrand
gaus<-function(x,alpha){
       # x shape is (batch, dim)
       # alpha is scalar - note the alpha[1] below
       # Compute sum of squared differences from 0.5
       dx2 = apply((x-0.5)**2,1,sum)
       return(exp(-alpha[1] * dx2))
 }
 # check function will work with matrix argument and return vector
 myx<-matrix(data=rep(0.7,10),ncol=5) # 2 row 5 col
 alpha<-matrix(data=c(1.),nrow=1) # scalar stored in matrix
 print(gaus(x=myx,alpha=alpha)) # should return vector of length 2

 # compute Vegas estimate
 result2<-vegas_integrate(f=gaus,lower=rep(0,5), upper=rep(1,5),
 nitn_warm = 10, neval_warm = 10000, nitn = 5, neval = 10000,
 errTol=0.01,maxIter=100,
 alpha=alpha)
 print(result2)

 #compute known numerical result
 alpha<-matrix(data=c(1.),nrow=1)
 erf<-function(x){return(2 * pnorm(x * sqrt(2)) - 1)}
 oneD<-sqrt(pi / alpha[1]) * erf(sqrt(alpha[1]) / 2.0)
 true.val<-oneD**5
 print(true.val)
 cat("Vegas =",result2$mean," known estimate =",true.val,"\n")

} # }